Added add &  delete items from custom list and lodash

+++ app.js
@@ -3,6 +3,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 
@@ -107,7 +108,7 @@ app.get("/", function (req, res) {
 });
 
 app.get("/:customListName", function (req, res) {
-  const customListName = req.params.customListName;
+  const customListName = _.capitalize(req.params.customListName);
 
   List.findOne({ name: customListName }).then(function (foundList) {
     if (!foundList) {
@@ -132,21 +133,21 @@ app.get("/:customListName", function (req, res) {
 app.post("/", function (req, res) {
 
   const itemName = req.body.newItem;
-  const listName=req.body.list;
+  const listName = req.body.list;
 
   const item = new Item({
     name: itemName
   });
 
-  if (listName==="Today") {
+  if (listName === "Today") {
     item.save();
-    res.redirect("/"); 
+    res.redirect("/");
   }
-  else{
-    List.findOne({name:listName}).then(function (foundList) {
+  else {
+    List.findOne({ name: listName }).then(function (foundList) {
       foundList.items.push(item);
       foundList.save();
-      res.redirect("/"+listName);
+      res.redirect("/" + listName);
     })
   }
 
@@ -154,14 +155,25 @@ app.post("/", function (req, res) {
 
 app.post("/delete", function (req, res) {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId).then(function () {
-    console.log("Deleted");
-    res.redirect("/");
-  })
-    .catch(function (err) {
-      console.log(err);
-    });
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId).then(function () {
+      console.log("Deleted");
+      res.redirect("/");
+    })
+      .catch(function (err) {
+        console.log(err);
+      });
+  }
+  else {
+    List.findOneAndUpdate({ name: listName }, { $pull: { items: { _id: checkedItemId } } }).then(function (foundList) {
+      res.redirect("/" + listName);
+    })
+      .catch(function (err) {
+        console.log(err);
+      })
+  }
 
 
 })
